関数の定義を共通化
