+++ 03.asynchronous/async.js
@@ -1,41 +1,7 @@
-import sqlite3 from "sqlite3";
 import timers from "timers/promises";
-const db = new sqlite3.Database(":memory:");
-
-const runSql = (sql) =>
-  new Promise((resolve, reject) => {
-    db.run(sql, (error) => {
-      if (error) {
-        reject(error);
-      } else {
-        resolve();
-      }
-    });
-  });
-
-const runSqlToInsert = (sql) =>
-  new Promise((resolve, reject) => {
-    db.run(sql, function (error) {
-      if (error) {
-        reject(error);
-      } else {
-        console.log(`ID${this.lastID}が挿入されました`);
-        resolve(`ID${this.lastID}が挿入されました`);
-      }
-    });
-  });
-
-const displayAll = (sql) =>
-  new Promise((resolve, reject) => {
-    db.all(sql, (error, rows) => {
-      if (error) {
-        reject(error);
-      } else {
-        rows.forEach((element) => console.log(element));
-        resolve();
-      }
-    });
-  });
+import { runSql } from "./function.js";
+import { runSqlToInsert } from "./function.js";
+import { displayAll } from "./function.js";
 
 //エラーなし
 
@@ -43,8 +9,14 @@ async function sequentialStart() {
   await runSql(
     "CREATE TABLE books (id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, title VARCHAR(10) UNIQUE)"
   );
-  await runSqlToInsert('INSERT INTO books(title) VALUES("チェリー本")');
-  await runSqlToInsert('INSERT INTO books(title) VALUES("ブルーベリー本")');
+  let result = await runSqlToInsert(
+    'INSERT INTO books(title) VALUES("チェリー本")'
+  );
+  console.log(result);
+  let result2 = await runSqlToInsert(
+    'INSERT INTO books(title) VALUES("ブルーベリー本")'
+  );
+  console.log(result2);
   await displayAll("SELECT * FROM books");
   runSql("drop table if exists books");
 }
@@ -53,7 +25,7 @@ sequentialStart();
 
 await timers.setTimeout(100);
 
-// エラーあり
+// // エラーあり
 async function errorSequential() {
   await runSql(
     "CREATE TABLE books (id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, title VARCHAR(10) UNIQUE)"
